refactor: update folder and file structure
